% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayat.R
\name{ayat}
\alias{ayat}
\title{Ambil satu ayat dari Al-Qur'an}
\usage{
ayat(nomor_surat, nomor_ayat)
}
\arguments{
\item{nomor_surat}{Nomor surat, dari 1 sampai 114.}

\item{nomor_ayat}{Nomor ayat, berupa satu angka atau beberapa angka.}
}
\value{
Data frame dengan satu baris berisi nomor surat,
nomor ayat, dan terjemahan Indonesia.
}
\description{
Mengembalikan satu ayat berdasarkan nomor surat dan nomor ayat.
}
\examples{
ayat(1, 1)
ayat(2, 255)
ayat(114, 1:3)
ayat(2, c(255, 256))

}
