% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cari.R
\name{cari}
\alias{cari}
\title{Cari kata atau frasa dalam terjemahan Al-Qur'an}
\usage{
cari(kata)
}
\arguments{
\item{kata}{Kata atau frasa yang ingin dicari.}
}
\value{
Data frame berisi surat, ayat, dan terjemahan
yang mengandung kata atau frasa tersebut.
}
\description{
Mencari kata atau frasa dalam terjemahan Al-Qur'an
Bahasa Indonesia.
}
\examples{
cari("Allah")
cari("rahmat")
cari("orang-orang")

}
