% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surat.R
\name{surat}
\alias{surat}
\title{Ambil satu surat dari Al-Qur'an}
\usage{
surat(nomor)
}
\arguments{
\item{nomor}{Nomor surat, dari 1 sampai 114.}
}
\value{
Data frame berisi nomor surat, nomor ayat,
teks Arab, dan terjemahan Indonesia.
}
\description{
Mengembalikan seluruh ayat dari surat yang dipilih,
beserta teks Arab dan terjemahan Bahasa Indonesia.
}
\examples{
surat(1)
surat(2)
surat(114)

}
