% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tulis_ayat_fragmen.R
\name{tulis_ayat_fragmen}
\alias{tulis_ayat_fragmen}
\title{Menulis fragmen ayat Al-Qur'an untuk Quarto}
\usage{
tulis_ayat_fragmen(surat, ayat, bagian = NULL)
}
\arguments{
\item{surat}{Nomor surat Al-Qur'an.}

\item{ayat}{Nomor ayat.}

\item{bagian}{Nomor bagian ayat yang ingin ditampilkan.
Jika \code{NULL}, semua bagian ditampilkan.}
}
\value{
Objek HTML yang dapat dirender oleh Quarto.
}
\description{
Menampilkan satu atau beberapa bagian dari teks Arab ayat
berdasarkan tanda waqaf.
}
\examples{
if (interactive()) {
  tulis_ayat_fragmen(2, 2)
  tulis_ayat_fragmen(2, 286, 7)
  tulis_ayat_fragmen(2, 286, 3:5)
}
}
