% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayat.R
\name{ayat}
\alias{ayat}
\title{Mengambil teks ayat Al-Qur'an}
\usage{
ayat(nomor_surat, nomor_ayat)
}
\arguments{
\item{nomor_surat}{Nomor surat Al-Qur'an.}

\item{nomor_ayat}{Nomor ayat atau beberapa nomor ayat.}
}
\value{
Data frame berisi nomor surat, nomor ayat, teks Arab,
dan terjemahan bahasa Indonesia.
}
\description{
Mengambil teks Arab dan terjemahan bahasa Indonesia berdasarkan
nomor surat dan nomor ayat.
}
